refactor: mv getProtocol
